+++ ArcdpsLogManager/CHANGELOG.md
@@ -5,6 +5,7 @@ This is the full changelog of the arcdps Log Manager.
 ## Log Manager v1.11
 
 #### New features
+- Added a "Weekly clears" tab with a table of encounters finished since weekly reset
 - Added Cosmic Observatory CM detection
 - Added Temple of Febe CM detection